Solution html 요소 컴퍼넌트 단위로 분리
